+++ ui/js/spotlight.js
@@ -1,4 +1,4 @@
-import { mk, clr, el, inp as inpById, q, qa, setUserText } from '/js/utils.js?v=0cfffdfb';
+import { mk, clr, el, inp as inpById, q, qa, setUserText } from '/js/utils.js?v=2e617767';
 import { wrapTab } from '/js/dialog.js?v=05935547';
 import { t } from '/js/i18n.js?v=d056c9c5';